# backend/backend/docs/swagger/messenger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Suda Messenger: authentication, chats, messaging, users, search.",
        "title": "Suda Messenger Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/chats/{id}/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search messages within a specific chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search query (min 2 chars)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.ChatSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search users, chats and messages globally",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (min 2 chars)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.GlobalSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.ChatSearchResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.SearchResult"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.GlobalSearchResponse": {
            "type": "object",
            "properties": {
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.SearchResult"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.SearchResult"
                    }
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.SearchResult"
                    }
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_search.SearchResult": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "highlight": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "message_id": {
                    "type": "integer"
                },
                "rank": {
                    "type": "number"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}